Move redux folder to src
